// src/interfaces/IDashboardPage.ts
import { Locator } from "@playwright/test";

export interface IDashboardPageLocators {
    getAutomation_hdr: Locator;
    getSubHeader_txt: Locator;
    getHomeBtn: Locator;
    getOrdersBtn: Locator;
    getCartBtn: Locator;
    getProductCard: Locator;
    getProduct_title: Locator;
    getBlinkingLabel:Locator;
    getCardCount:Locator;
}